+++ TestSuite/pack.py
@@ -44,6 +44,7 @@ class Pack:
         self.generic_fields: List[JSONBased] = list()
         self.generic_types: List[JSONBased] = list()
         self.generic_modules: List[JSONBased] = list()
+        self.generic_definitions: List[JSONBased] = list()
         self.layouts: List[JSONBased] = list()
         self.layoutcontainers: List[JSONBased] = list()
         self.reports: List[JSONBased] = list()
@@ -51,7 +52,6 @@ class Pack:
         self.playbooks: List[Playbook] = list()
         self.test_playbooks: List[Playbook] = list()
         self.release_notes: List[TextBased] = list()
-
         # Create base pack
         self._pack_path = packs_dir / name
         self._pack_path.mkdir()
@@ -98,6 +98,9 @@ class Pack:
 
         self._generic_modules_path = self._pack_path / 'GenericModules'
         self._generic_modules_path.mkdir()
+
+        self._generic_definitions_path = self._pack_path / 'GenericDefinitions'
+        self._generic_definitions_path.mkdir()
 
         self._layout_path = self._pack_path / 'Layouts'
         self._layout_path.mkdir()
@@ -284,8 +287,10 @@ class Pack:
             self,
             name,
             content: dict = None) -> JSONBased:
+        dir_path = self._generic_fields_path / name
+        dir_path.mkdir()
         prefix = 'genericfield'
-        generic_field = self._create_json_based(name, prefix, content, dir_path=self._generic_fields_path)
+        generic_field = self._create_json_based(name, prefix, content, dir_path=dir_path)
         self.generic_fields.append(generic_field)
         return generic_field
 
@@ -293,8 +298,10 @@ class Pack:
             self,
             name,
             content: dict = None) -> JSONBased:
+        dir_path = self._generic_types_path / name
+        dir_path.mkdir()
         prefix = 'generictype'
-        generic_type = self._create_json_based(name, prefix, content, dir_path=self._generic_types_path)
+        generic_type = self._create_json_based(name, prefix, content, dir_path=dir_path)
         self.generic_types.append(generic_type)
         return generic_type
 
@@ -306,6 +313,15 @@ class Pack:
         generic_module = self._create_json_based(name, prefix, content, dir_path=self._generic_modules_path)
         self.generic_modules.append(generic_module)
         return generic_module
+
+    def create_generic_definition(
+            self,
+            name,
+            content: dict = None) -> JSONBased:
+        prefix = 'genericdefinition'
+        generic_definition = self._create_json_based(name, prefix, content, dir_path=self._generic_definitions_path)
+        self.generic_definitions.append(generic_definition)
+        return generic_definition
 
     def create_layout(
             self,